https://www.codewars.com/kata/decimal-decomposition

+++ kata/7 kyu/decimal-decomposition/main/DecimalDecomposition.java
@@ -1,6 +1,6 @@
 class DecimalDecomposition {
   static String decimalDecomposition(int number) {
-    int i = (("" + number).charAt(0) - '0') * (int) Math.pow(10, Math.log10(number));
+    int i = (("" + number).charAt(0) - '0') * (int) Math.pow(10, (int) Math.log10(number));
     return number == i ? i + "" : i + "+" + decimalDecomposition(number - i);
   }
 }
